pyproject.toml
[tool.poetry]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI application with PostgreSQL"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "app", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.0.3"
sqlmodel = "^0.0.14"
requests = "^2.32.3"
asyncpg = "^0.30.0"
aiohttp = "^3.11.13"
alembic = "^1.15.1"
isort = "^6.0.1"
pytest-asyncio = "^0.25.3"
aiosqlite = "^0.21.0"
greenlet = "^3.1.1"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"] 